Add files via upload
This is the .C file that performs the same computation as the python script.
Note that cpu performance data is not directly available from the code itself. This data must be obtained indirectly from the debugger tools or the 
built-in functions on the terminal. 

Program timing is supported on both functions.