Create Code_Solutions
This is the final code used to gather initial MPI usage data for both serial and parallel algorithms. The case consideration problem is worded as follows:
"""
By considering the terms in the Fibonacci sequence whose values do not exceed four million, find the sum of the even-valued terms. Compute this result 
exactly 100 times.
"""
From this problem statement, the C++ code with mpi was developed. Select the appropriate file to run the mpi solution.